+++ pytorch_lightning/plugins/native_amp.py
@@ -67,8 +67,6 @@ class NativeAMPPlugin(PrecisionPlugin):
             torch.nn.utils.clip_grad_value_(model.parameters(), clip_value=grad_clip_val)
         elif gradient_clip_algorithm.startswith('norm'):
             torch.nn.utils.clip_grad_norm_(model.parameters(), max_norm=grad_clip_val, norm_type=norm_type)
-        else:
-            raise ValueError(f'gradient_clip_algorithm [{gradient_clip_algorithm}] is not valid.')
 
     @property
     def scaler(self):